+++ packages/headless/src/controllers/facets/range-facet/date-facet/headless-date-facet.ts
@@ -1,11 +1,11 @@
+import dayjs from 'dayjs';
+import utc from 'dayjs/plugin/utc';
+import customParseFormat from 'dayjs/plugin/customParseFormat';
+
 import {
   DateRangeRequest,
   DateFacetRequest,
 } from '../../../../features/facets/range-facets/date-facet-set/interfaces/request';
-import {
-  AutomaticRangeFacetOptions,
-  ManualRangeFacetOptions,
-} from '../../../../features/facets/range-facets/generic/interfaces/options';
 import {Engine} from '../../../../app/headless-engine';
 import {randomID} from '../../../../utils/utils';
 import {DateFacetRegistrationOptions} from '../../../../features/facets/range-facets/date-facet-set/interfaces/options';
@@ -21,26 +21,51 @@ import {
   SearchSection,
 } from '../../../../state/state-sections';
 import {executeToggleDateFacetSelect} from '../../../../features/facets/range-facets/date-facet-set/date-facet-controller-actions';
+import {validateOptions} from '../../../../utils/validate-payload';
+import {
+  DateFacetOptions,
+  dateFacetOptionsSchema,
+} from './headless-date-facet-options';
 
-type DateRangeOptions = Pick<DateRangeRequest, 'start' | 'end'> &
-  Partial<DateRangeRequest>;
+dayjs.extend(utc);
+dayjs.extend(customParseFormat);
+
+type DateRangeOptions = Partial<Omit<DateRangeRequest, 'start' | 'end'>> & {
+  start: string | number | Date;
+  end: string | number | Date;
+  useLocalTime?: boolean;
+  dateFormat?: string;
+};
 
 export function buildDateRange(config: DateRangeOptions): DateRangeRequest {
+  const DATE_FORMAT = 'YYYY/MM/DD@HH:mm:ss';
+  const start = config.useLocalTime
+    ? dayjs(config.start, config.dateFormat).format(DATE_FORMAT)
+    : dayjs(config.start, config.dateFormat).utc().format(DATE_FORMAT);
+  const end = config.useLocalTime
+    ? dayjs(config.end, config.dateFormat).format(DATE_FORMAT)
+    : dayjs(config.end, config.dateFormat).utc().format(DATE_FORMAT);
+
+  if (start === 'Invalid Date' || end === 'Invalid Date') {
+    throw new Error(
+      `Could not parse the provided date, please provide a dateFormat string in the configuration options.\n
+       See https://day.js.org/docs/en/parse/string-format for more information.
+       `
+    );
+  }
   return {
     endInclusive: false,
     state: 'idle',
     ...config,
+    start,
+    end,
   };
 }
 
+export {DateFacetOptions};
 export type DateFacetProps = {
   options: DateFacetOptions;
 };
-
-export type DateFacetOptions = {facetId?: string} & (
-  | Omit<AutomaticRangeFacetOptions<DateFacetRequest>, 'facetId'>
-  | Omit<ManualRangeFacetOptions<DateFacetRequest>, 'facetId'>
-);
 
 /** The `DateFacet` controller makes it possible to create a facet with date ranges.*/
 export type DateFacet = ReturnType<typeof buildDateFacet>;
@@ -54,6 +79,8 @@ export function buildDateFacet(
 
   const facetId = props.options.facetId || randomID('dateFacet');
   const options: DateFacetRegistrationOptions = {facetId, ...props.options};
+
+  validateOptions(dateFacetOptionsSchema, options, buildDateFacet.name);
 
   dispatch(registerDateFacet(options));