+++ packages/headless/src/controllers/facets/range-facet/numeric-facet/headless-numeric-facet.test.ts
@@ -49,6 +49,13 @@ describe('numeric facet', () => {
     expect(engine.actions).toContainEqual(action);
   });
 
+  it('when an option is invalid, it throws an error', () => {
+    options.numberOfValues = 0;
+    expect(() => initNumericFacet()).toThrow(
+      'Check the options of buildNumericFacet'
+    );
+  });
+
   describe('#toggleSelect', () => {
     it('dispatches a toggleSelectNumericFacetValue with the passed value', () => {
       const value = buildMockNumericFacetValue();