+++ packages/headless/src/features/facets/range-facets/generic/range-facet-controller-actions.ts
@@ -5,6 +5,13 @@ import {ConfigurationSection} from '../../../../state/state-sections';
 import {getAnalyticsActionForToggleRangeFacetSelect} from './range-facet-utils';
 import {updateFacetOptions} from '../../../facet-options/facet-options-actions';
 import {executeSearch} from '../../../search/search-actions';
+import {validatePayloadSchema} from '../../../../utils/validate-payload';
+import {RecordValue} from '@coveo/bueno';
+import {facetIdDefinition} from '../../generic/facet-actions-validation';
+import {
+  numericFacetValueDefinition,
+  dateFacetValueDefinition,
+} from './range-facet-validate-payload';
 
 /**
  * Executes a search with the appropriate analytics for a toggle range facet value
@@ -19,6 +26,11 @@ export const executeToggleRangeFacetSelect = createAsyncThunk<
   },
   AsyncThunkSearchOptions<ConfigurationSection>
 >('rangeFacet/executeToggleSelect', ({facetId, selection}, {dispatch}) => {
+  validatePayloadSchema(
+    {facetId, selection},
+    {facetId: facetIdDefinition, selection: getSelectionDefinition(selection)}
+  );
+
   const analyticsAction = getAnalyticsActionForToggleRangeFacetSelect(
     facetId,
     selection
@@ -27,3 +39,10 @@ export const executeToggleRangeFacetSelect = createAsyncThunk<
   dispatch(updateFacetOptions({freezeFacetOrder: true}));
   dispatch(executeSearch(analyticsAction));
 });
+
+const getSelectionDefinition = (selection: RangeFacetValue) => {
+  if (typeof selection.start === 'string') {
+    return new RecordValue({values: dateFacetValueDefinition});
+  }
+  return new RecordValue({values: numericFacetValueDefinition});
+};