+++ packages/headless/src/integration-tests/search-app.test.ts
@@ -1,4 +1,9 @@
 import {
+  buildDateSortCriterion,
+  buildRelevanceSortCriterion,
+  SortOrder,
+} from '../features/sort-criteria/criteria';
+import {
   HeadlessEngine,
   searchAppReducers,
   buildSearchBox,
@@ -9,8 +14,6 @@ import {
   SearchActions,
   AnalyticsActions,
   Result,
-  buildRelevanceSortCriterion,
-  buildDateSortCriterion,
   CategoryFacetValue,
   FacetValue,
 } from '../index';
@@ -87,7 +90,7 @@ describe('search app', () => {
 
     beforeAll(async () => {
       initialResults = resultList.state.results;
-      sort.sortBy(buildDateSortCriterion('descending'));
+      sort.sortBy(buildDateSortCriterion(SortOrder.Descending));
 
       await sleep(2);
     });